Graph Framework: Introduce new plugin for graph display handling
Topology editor feature is coming. For now, we only had to display
an existing topology. To do that, we use an FDK dedicated displayer
which is tightly related to FDK Modules model.

The problem of this implementation is that we are forced to match this
representation even when it is not the case. For example, during debug
topology display, we create partial modules to display the topology.

This patch introduces a plugin which aims to display a simple graph
structure. A graph is composed by nodes and edges. Each node can have
several entry/exit point where edges can be connected.
This plugin can be extended with new controllers which are able to
custom nodes/edges representation.

Tracked-On: RTC-173279
Change-Id: If3325dc25ade1ef210c4cc5b517a90b5a15d6cca
Signed-off-by: Jules Clero <julesx.clero@intel.com>