Graph Framework: Add a Drag and Drop Controller for the Graph displayer
This controller allows to detect drop events of a particular type of
object on the displayer.
Clients can listen for events providing the dropped object and the
graph element on which it has been dropped.
For now, the target graph element is not detected, the returned parent
is always the root Graph.

Tracked-On: RTC-174221
Change-Id: Ia4548dba372f50ea9cda6ae64cfc7e03b1ed35b2
Signed-off-by: Jules Clero <julesx.clero@intel.com>