Graph Framework: Force presenters to layout sub-figures at creation time
Previous implementation was adding listeners to put pins figure of a
LeafPresenter at their right place. Nevertheless, this listener was
trigged only after the first location update of the figure.
Thus, the calculated width in the AutoLayoutGenerator is not the
presenter width but only the NodeBodyFigure one.
Moreover, this bug leads to the display of a figure with pin at the
wrong location in the topology editor (once again pins are placed only
after the first move).

This patch fix the bug by launching the presenter algorithm once in
constructor (location listeners remains too).

Tracked-On: RTC-174221
Change-Id: I5401491776c0554e3035b749e72af9e82f0bc885
Signed-off-by: Jules Clero <julesx.clero@intel.com>