Graph Framework: Fix GroupPresenter to be moved in editor
This patch fixes two problems. First the GroupPresenter cannot be moved
when it is empty, this problem is fixed by updating the bounds figure
location even when no sub-presenter is present.
Secondly, the implementation was buggy and sub-presenters added after
the construction was not taken into account when updating the group
bounds figure.

Tracked-On: RTC-175985
Change-Id: Ie111cbc4a8d99ccfd004110a40fe7046d5520e67
Signed-off-by: Jules Clero <julesx.clero@intel.com>