Selection Controller: Display context menu on mouse release events
This patch move the context menu display on the mouse release events.
It fix a bug of the node move controller. The ghost may not be removed
when the menu is displayed, indeed the mouse release event is captured
by the menu. This patch allows to let all controllers of the event
before the menu display.
Moreover it matches the behavior of several other tools.

Change-Id: I5977113d11450a465c2e3f4dd503a2ccfdc855b8
Signed-off-by: Jules Clero <julesx.clero@intel.com>