Topology editors: Use hot point concept to enhance user experience
This patch adds the support of the hot point concept in both
the test and the linux topology editors.
The DND and the reparenting controller have been updated to inform
listeners of the event location.
This information is transmitted to the layout generator which use
it as hot point. Thus, newly created or reparented node presenter
do not any more appears at position (0, 0) but rather where the event
occurs.

Change-Id: I8103f252e14a181bad9d522481a2935a492ffb4e
Signed-off-by: Jules Clero <julesx.clero@intel.com>