Graph Framework: Allow to set location of empty groups
Current auto layout implementation does give a position
to a group presenter. Indeed, the position is automatically
updated when a child leaf presenter position is updated.
This behavior does not work when dealing with empty groups.
This patch fix the problem by allowing the AutoLayoutGenerator to
retrieve empty group calculated position.

Tracked-On: RTC-183760
Change-Id: Ie4a74187c49a3e09f8918f8909bd63f55a0590e2
Signed-off-by: Jules Clero <julesx.clero@intel.com>