Graph Framework: Fix circular cycle graph auto layout
Current implementation of the auto layout algorithm
fails to layout graph which have cycle and no "Source node".
Indeed, the graph algorithm is based on the discovery
of nodes used as "Source" which can be inputless or which
does not have any connected input.
Such nodes may not be present in circular graph.
This patch fix the algorithm in order to be able to layout
those graphs.

Tracked-On: RTC-183851
Change-Id: I7b1e6d3dbf3068dc7930dfd9a445fd91fc9013a2
Signed-off-by: Jules Clero <julesx.clero@intel.com>