Graph Framework: Retrieve figure bg/fg colors from element style
This patch add the foreground and background colors to Graph element
style object.
Colors are now retrieved from those objects.
This patch requires that all figures are linked to the model element
they represent.This is not the case for ghost figures used in the
link controller. This bug will be fixed in later patches.

Tracked-On: RTC-182516
Change-Id: Iea386f0515aa5c95eb60e874ed2e45d1f654aa9b
Signed-off-by: Jules Clero <julesx.clero@intel.com>